Library API

Contents

Library API

Contents

Index

GSB Signal

SE Signal

ESA Signal

Constructors

Misc.

Deprecated

Index

WaveMixings.cite

WaveMixings.esa

WaveMixings.esa_dispersion

WaveMixings.esa_lgc

WaveMixings.esa_strong

WaveMixings.fs2toau2

WaveMixings.fs3toau3

WaveMixings.gsb

WaveMixings.gsb_dispersion

WaveMixings.gsb_strong

WaveMixings.make_door_window!

WaveMixings.make_nth_order_dispersion

WaveMixings.make_probe_freq

WaveMixings.make_pulse_duration

WaveMixings.make_pump_amplitude

WaveMixings.make_pump_freq

WaveMixings.pkginfo

WaveMixings.print_logo

WaveMixings.se

WaveMixings.se_dispersion_lgc

WaveMixings.se_fluorescence!

WaveMixings.se_lgc

WaveMixings.se_strong_lgc

WaveMixings.Δgsb

WaveMixings.Δse

WaveMixings.Δse_lgc

WaveMixings.Δse_strong_lgc

GSB Signal

WaveMixings.gsb — Function

gsb(pump::T, Τ::T, self::WaveMixing, frame::Integer=1)  where {T <: Real}

Calculates the signal from energies and oscillator strengths as input. The inputs must be 1D Array, i.e. vectors. The signal is calculated from the spectrum energy in the frequency domain. For the available methods check the function spectrum_energy_fourier. Use with Ground State Dynamics.

Arguments

pump: pump pulse energy.

Τ: pulse length.

WaveMixing: composite containing the method parameter. method by default is Gaussian (:gaus). Lorentzian (:lore), chirp (:chir), hyperbolic secant (:sech), and asymmetric hyperbolic secant (:assh) are available.

frame: frame number in the non-adiabatic trajectory. The default gives 1.

Returns

GSB door, single value.

Description

This function estimates the Doorway (door) parameter under the following approximations:

Evolutions of electronic coherences in the DW operators are evaluated within the short-time classical Condon approximation.

T-evolutions of electronic populations, are replaced by evolutions along trajectories.

The quantum thermal distribution in the electronic ground state is substituted by the corresponding phase-space Wigner distribution.

The trace over vibrational degrees of freedom is replaced by stochastic sampling of the initial conditions.

The corresponding Doorway formula (see Ref. [1] formula 56) of the ground state bleach (GSB) reads:

\[D_{0}(\omega_{pu},\boldsymbol{R}_{g},\boldsymbol{P}_{g}) = \sum_{e}^{ } E_{pu}^{2}(\omega_{pu}-U_{eg}(\boldsymbol{R}_{g})) \left| (\boldsymbol{\mu}_{ge}(\boldsymbol{R}_{g}) \right|^{2} \rho_{g}^{\textrm{Wig}}(\boldsymbol{R}_{g},\boldsymbol{P}_{g})\]

Cite

[1] Gelin et.al. J.Chem.TheoryComput.2021, 17, 2394−2408

See also gsb, and gsb_strong.

gsb(probe::AbstractMatrix{T}, Τ::T, self::WaveMixing) where {T <: Real}

Calculates the signal from energies and oscillator strengths as input. The inputs must be 2D Array, i.e. Matrix. The signal is calculated from the spectrum energy in the frequency domain. For the available methods check the function spectrum_energy_fourier. Use with Ground State Dynamics.

Arguments

probe: probe pulse energies.

Τ: pulse length.

WaveMixing: composite containing oscillators, method and energies parameters.

Returns

Calculates window and updates the WaveMixing composite (GS) with it. (Matrix with n-probe columns, one column has all sums for one probe pulse energy).

Description

This function estimates the Window (window) parameter under the following approximations:

Evolutions of electronic coherences in the DW operators are evaluated within the short-time classical Condon approximation.

T-evolutions of electronic populations, are replaced by evolutions along trajectories.

The quantum thermal distribution in the electronic ground state is substituted by the corresponding phase-space Wigner distribution.

The trace over vibrational degrees of freedom is replaced by stochastic sampling of the initial conditions.

The corresponding Window formula (see Ref. [1] formula 58) of the ground state bleach (GSB) reads:

\[W_{0}^{int}(\omega_{pr},\boldsymbol{R}_{g},\boldsymbol{P}_{g}) = \sum_{e}^{ } E_{pr}^{2}(\omega_{pr}-U_{eg}(\boldsymbol{R}_{g})) \left| (\boldsymbol{\mu}_{ge}(\boldsymbol{R}_{g}) \right|^{2}\]

Cite

[1] Gelin et.al. J.Chem.TheoryComput.2021, 17, 2394−2408

See also gsb, and gsb_strong.

WaveMixings.gsb_dispersion — Function

gsb_dispersion(probe::T, dispersion::AbstractMatrix{T}, Τ::T,
               self::WaveMixing, dephase::T=0.01) where {T <: Real}

Calculates the dispersed signal (eq. 62 in Gelin JCTC, 2394. 2021) from energies and oscillator strengths as input. The inputs must be 2D Array, i.e. Matrix. The signal is calculated from the spectrum energy in the frequency domain. For the available methods check the function spectrum_energy_fourier. Use with Ground State Dynamics.

Arguments

probe: a probe pulse energy.

dispersion: array of dispersion energies.

Τ: pulse length.

WaveMixing: composite containing oscillators, method and energies parameters.

dephase: dephasing parameter. the default gives 66 fs.

Returns

Calculates dispersed window and updates the WaveMixing composite (GS) with it. (Matrix with n-probe columns, one column has all sums for one probe pulse energy).

Description

This function estimates the dispersed Window (window) transient-absorption pump-probe signal, which derives from performing the Fourier transform to the integral pump-probe signal (see gsb) with respect to $\tau_{t}$. The dispersed pump-probe signal is sensitive to electronic dephasing ($\lambda$), a parameter of this function.

The corresponding dispersed Window formula (see Ref. [1] formula 62) of the ground state bleach (GSB) reads:

\[W_{0}^{dis}(\omega_{pr},\omega,\boldsymbol{R}_{g}(T),\boldsymbol{P}_{g}(T)) = E_{pr}^{2}(\omega-\omega_{pr}) \sum_{e}\frac{\nu}{\nu^{2}+(\omega-\omega_{eg}(\boldsymbol{R}_{g}(T)))^{2}} |\mu_{ge}(\boldsymbol{R}_{g}(T))|^{2}\]

Cite

[1] Gelin et.al. J.Chem. Theory Comput.2021, 17, 2394−2408

See also se_dispersion and esa_dispersion.

WaveMixings.gsb_strong — Function

gsb_strong(pump::T, Τ::T, self::WaveMixing, frame::Integer=1, ) where {T <: Real}

Calculates the strong field signal from energies, oscillator strengths, and the pump laser amplitude as input.The inputs must be 1D Array, i.e. vectors. Use with Ground State Dynamics.

Arguments

pump: pump pulse energy.

Τ: pulse length.

WaveMixing: composite containing the λ parameter. λ is the pump laser amplitude.

frame: frame number in the non-adiabatic trajectory. The default gives 1.

Returns

Strong field GSB door, single value.

Description

This function estimates the strong field (door) parameter under the approximations of the Doorway-Window representation. The corresponding formula (see Ref. [1] formula XX) of the ground state bleach (GSB) reads:

\[ D_{0,\alpha}^{strong}(\omega_{\tau},\boldsymbol{R}_{g},\boldsymbol{P}_{g}) = \sum_{e'}\frac{ \lambda_{\textrm{pu}}^{2} \left| \mu_{\textrm{ge}}(\boldsymbol{R}_{\textrm{g}})\right|^{2}}{ \Omega_{\textrm{eg}}^{2} (\boldsymbol{R}_{g})} \left(1 - \cos\left(\Omega_{\textrm{eg}} \left(\boldsymbol{R}_{g}\right) \tau_{\textrm{pu}}\right) \right) \rho_{\textrm{g}}^{\textrm{Wig}} (\boldsymbol{R}_{g},\boldsymbol{P}_{g})\]

Cite

[1] Hefen et.al. XX 2025, XX

See also Δgsb_strong, Δse_strong, se_strong, and esa_strong.

gsb_strong(probe::AbstractMatrix{T}, Τ::T, self::WaveMixing) where {T <: Real}

Calculates the strong field Window from energies, oscillator strengths, and the probe laser amplitude as input.The inputs must be 1D Array, i.e. vectors. Use with Ground State Dynamics.

Arguments

probe: probe pulse energy.

Τ: pulse length.

WaveMixing: composite containing the λ parameter. λ is the pump laser amplitude.

Returns

Strong field GSB Window, matrix.

Description

This function estimates the strong field (window) parameter under the approximations of the Doorway-Window representation. The corresponding formula (see Ref. [1] formula XX) of the ground state bleach (GSB) reads:

\[\]

Cite

[1] XX et.al. J. Chem. Phys.XXXX, X, XXXX

See also Δgsb_strong, Δse_strong, se_strong, and esa_strong.

Missing docstring.

Missing docstring for gsb_wind_2d. Check Documenter's build log for details.

WaveMixings.Δgsb — Function

Δgsb(self::WaveMixing, frame::Integer=1, )

Calculates the Δ signal from energies and oscillator strengths as input. The inputs must be 1D Array, i.e. vectors. The signal is calculated merely from the dipole moments. Use with Ground State Dynamics.

Arguments

WaveMixing: composite containing oscillators, method and energies parameters.

frame: frame number in the non-adiabatic trajectory. The default gives 1.

Returns

GSB delta_door. Single value.

Description

This function estimates the Doorway (delta_door) parameter under the assumsion that the pump and probe pulses are well temporally separated i.e. the time delay (T) between the pulses is longer than the pulse durations, $\tau > T_{a}$ (see Ref. [1] formula 52)

\[D_{0}(\omega_{pu},\boldsymbol{R}_{g},\boldsymbol{P}_{g}) = \sum_{e}^{} \left| (\boldsymbol{\mu}_{ge}(\boldsymbol{R}_{g}) \right|^{2} \rho_{g}^{\textrm {Wig}}(\boldsymbol{R}_{g},\boldsymbol{P}_{g})\]

Cite

[1] Gelin et.al. J.Chem.TheoryComput.2021, 17, 2394−2408

See also Δgsb, and Δgsb_strong.

SE Signal

WaveMixings.se — Function

se(pump::T, Τ::T, self::WaveMixing, frame::Integer=1 ) where {T <: Real}

Calculates the SE Doorway from energies and oscillator strengths as input. The inputs must be 1D Array, i.e. vectors. The signal is calculated from the spectrum energy in the frequency domain. For the available methods check the function spectrum_energy_fourier. Use with Excited State Dynamics.

Arguments

pump: pump pulse energy.

Τ: pulse length.

WaveMixing: composite containing the method parameter. method by default is Gaussian (:gaus). Lorentzian (:lore), chirp (:chir), hyperbolic secant (:sech), and asymmetric hyperbolic secant (:assh) are available.

frame: frame number in the non-adiabatic trajectory. The default gives 1.

Returns

SE door. Single value.

Description

This function estimates the Doorway (door) parameter under the following approximations:

Evolutions of electronic coherences in the DW operators are evaluated within the short-time classical Condon approximation.

T-evolutions of electronic populations, are replaced by evolutions along trajectories.

The quantum thermal distribution in the electronic ground state is substituted by the corresponding phase-space Wigner distribution.

The trace over vibrational degrees of freedom is replaced by stochastic sampling of the initial conditions.

The corresponding Doorway formula (see Ref. [1] formula 57) of the stimulated emission (SE) reads:

\[D_{I}(\omega_{pu},\boldsymbol{R}_{g},\boldsymbol{P}_{g}) = E_{pu}^{2}(\omega_{pu}-U_{eg}(\boldsymbol{R}_{g})) \left| (\boldsymbol{\mu}_{ge}(\boldsymbol{R}_{g}) \right|^{2} \rho_{g}^{\textrm{Wig}}(\boldsymbol{R}_{g},\boldsymbol{P}_{g})\]

Cite

[1] Gelin et.al. J.Chem.TheoryComput.2021, 17, 2394−2408

See also se, and se_strong.

se(probe::AbstractMatrix{T}, Τ::T, self::WaveMixing) where {T <: Real}

Calculates the signal from energies and oscillator strengths as input. The inputs must be 2D Array, i.e. Matrix. The signal is calculated from the spectrum energy in the frequency domain. For the available methods check the function spectrum_energy_fourier. Use with Excited State Dynamics.

Arguments

probe: probe pulse energies.

Τ: pulse length.

WaveMixing: composite containing oscillators, method and energies parameters.

Returns

Calculates window and updates the WaveMixing composite (SE) with it. (Matrix with n-probe columns, one column has all sums for one probe pulse energy).

Description

This function estimates the Window (window) parameter under the following approximations:

Evolutions of electronic coherences in the DW operators are evaluated within the short-time classical Condon approximation.

T-evolutions of electronic populations, are replaced by evolutions along trajectories.

The quantum thermal distribution in the electronic ground state is substituted by the corresponding phase-space Wigner distribution.

The trace over vibrational degrees of freedom is replaced by stochastic sampling of the initial conditions.

The corresponding Window formula (see Ref. [1] formula 59) of the stimulated emission (SE) reads:

\[W_{I}^{int}(\omega_{pr},\boldsymbol{R}_{e},\boldsymbol{P}_{e}) = E_{pr}^{2}(\omega_{pr}-U_{e(T)g}(\boldsymbol{R}_{e})) \left| (\boldsymbol{\mu}_{ge(T)}(\boldsymbol{R}_{e}) \right|^{2} \]

Cite

[1] Gelin et.al. J.Chem.TheoryComput.2021, 17, 2394−2408

See also se, and se_strong.

WaveMixings.se_lgc — Function

se_lgc(se(pump::T, Τ::T, self::WaveMixing, frame::Integer=1 ) where {T <: Real}

LEGACY Calculates the SE Doorway from energies and oscillator strengths as input. The inputs must be 1D Array, i.e. vectors. The signal is calculated from the spectrum energy in the frequency domain. For the available methods check the function spectrum_energy_fourier. Use with Excited State Dynamics.

Arguments

pump: pump pulse energy.

Τ: pulse length.

WaveMixing: composite containing the method parameter. method by default is Gaussian (:gaus). Lorentzian (:lore), chirp (:chir), hyperbolic secant (:sech), and asymmetric hyperbolic secant (:assh) are available.

frame: frame number in the non-adiabatic trajectory. The default gives 1.

Returns

SE door. Single value.

Description

This function estimates the Doorway (door) parameter under the following approximations:

Evolutions of electronic coherences in the DW operators are evaluated within the short-time classical Condon approximation.

T-evolutions of electronic populations, are replaced by evolutions along trajectories.

The quantum thermal distribution in the electronic ground state is substituted by the corresponding phase-space Wigner distribution.

The trace over vibrational degrees of freedom is replaced by stochastic sampling of the initial conditions.

The corresponding Doorway formula (see Ref. [1] formula 57) of the stimulated emission (SE) reads:

\[D_{I}(\omega_{pu},\boldsymbol{R}_{g},\boldsymbol{P}_{g}) = E_{pu}^{2}(\omega_{pu}-U_{eg}(\boldsymbol{R}_{g})) \left| (\boldsymbol{\mu}_{ge}(\boldsymbol{R}_{g}) \right|^{2} \rho_{g}^{\textrm{Wig}}(\boldsymbol{R}_{g},\boldsymbol{P}_{g})\]

Cite

[1] Gelin et.al. J.Chem.TheoryComput.2021, 17, 2394−2408

See also se, and se_strong.

se_lgc(probe::AbstractMatrix{T}, Τ::T, self::WaveMixing) where {T <: Real}

LEGACY Calculates the signal from energies and oscillator strengths as input. The inputs must be 2D Array, i.e. Matrix. The signal is calculated from the spectrum energy in the frequency domain. For the available methods check the function spectrum_energy_fourier. Use with Excited State Dynamics.

Arguments

probe: probe pulse energies.

Τ: pulse length.

WaveMixing: composite containing oscillators, method and energies parameters.

Returns

Calculates window and updates the WaveMixing composite (SE) with it. (Matrix with n-probe columns, one column has all sums for one probe pulse energy).

Description

This function estimates the Window (window) parameter under the following approximations:

Evolutions of electronic coherences in the DW operators are evaluated within the short-time classical Condon approximation.

T-evolutions of electronic populations, are replaced by evolutions along trajectories.

The quantum thermal distribution in the electronic ground state is substituted by the corresponding phase-space Wigner distribution.

The trace over vibrational degrees of freedom is replaced by stochastic sampling of the initial conditions.

The corresponding Window formula (see Ref. [1] formula 59) of the stimulated emission (SE) reads:

\[W_{I}^{int}(\omega_{pr},\boldsymbol{R}_{e},\boldsymbol{P}_{e}) = E_{pr}^{2}(\omega_{pr}-U_{e(T)g}(\boldsymbol{R}_{e})) \left| (\boldsymbol{\mu}_{ge(T)}(\boldsymbol{R}_{e}) \right|^{2} \]

Cite

[1] Gelin et.al. J.Chem.TheoryComput.2021, 17, 2394−2408

See also se, and se_strong.

WaveMixings.se_dispersion_lgc — Function

se_dispersion_lgc(probe::T, dispersion::AbstractMatrix{T}, Τ::T,
                  self::WaveMixing, dephase::T=0.01 ) where {T <: Real}

LEGACY Calculates the dispersed signal (eq. 63 in Gelin JCTC, 2394. 2021) from energies and oscillator strengths as input. The inputs must be 2D Array, i.e. Matrix. The signal is calculated from the spectrum energy in the frequency domain. For the available methods check the function spectrum_energy_fourier. Use with Excited State Dynamics.

Arguments

probe: a probe pulse energy.

dispersion: array of dispersion energies.

WaveMixing: composite containing oscillators, method and energies parameters.

Τ: pulse length.

dephase: dephasing parameter. the default gives 66 fs.

Returns

Calculates dispersed window and updates the WaveMixing composite (SE) with it. (Matrix with n-probe columns, one column has all sums for one probe pulse energy).

Description

This function estimates the dispersed Window (window) transient-absorption pump-probe signal, which derives from performing the Fourier transform to the integral pump-probe signal (see se) with respect to $\tau_{t}$. The dispersed pump-probe signal is sensitive to electronic dephasing ($\lambda$), a parameter of this function.

The corresponding dispersed Window formula (see Ref. [1] formula 63) of the stimulated emission (SE) reads:

\[W_{\mathrm{I}}^{dis}(\omega_{pr},\omega,\boldsymbol{R}_{e}(T),\boldsymbol{P}_{e}(T)) = E_{pr}^{2}(\omega-\omega_{pr})\frac{\nu}{\nu^{2}+(\omega-\omega_{e(T)g}(\boldsymbol{R}_{e}(T)))^{2}} |\mu_{ge(T)}(\boldsymbol{R}_{e}(T))|^{2}\]

Cite

[1] Gelin et.al. J.Chem.TheoryComput.2021, 17, 2394−2408

See also gsb_dispersion and esa_dispersion.

WaveMixings.se_strong_lgc — Function

se_strong_lgc(pump::T, Τ::T, self::WaveMixing, frame::Integer=1) where {T <: Real}

LEGACY Calculates the strong field signal from energies, oscillator strengths, and the pump laser amplitude as input. The inputs must be 1D Array, i.e. vectors. The signal is calculated merely from the dipole moments. Use with Excited State Dynamics.

Arguments

pump: pump pulse energy.

Τ: pulse length.

WaveMixing: composite containing the λ parameter. λ is the pump laser amplitude.

frame: frame number in the non-adiabatic trajectory. The default gives 1.

Returns

Strong field SE door, single value.

Description

This function estimates the strong field (door) parameter under the approximations of the Doorway-Window representation. The corresponding formula (see Ref. [1] formula XX) of the stimulated emission (SE) reads:

\[ D_{I,\alpha}^{strong}(\omega_{\tau},\boldsymbol{R}_{e},\boldsymbol{P}_{e}) = \frac{ \lambda_{\textrm{pr}}^{2} \left| \mu_{\textrm{ge}}(\boldsymbol{R}_{\textrm{e}})\right|^{2}}{ \Omega_{\textrm{eg}}^{2} (\boldsymbol{R}_{e})} \left(1 - \cos\left(\Omega_{\textrm{eg}} \left(\boldsymbol{R}_{e}\right) \tau_{\textrm{pr}}\right) \right) (\boldsymbol{R}_{e},\boldsymbol{P}_{e})\]

Cite

[1] Hefen et.al. XX 2025, XX

See also Δgsb_strong, gsb_strong, Δse_strong, and esa_strong.

se_strong_lgc(probe::AbstractMatrix{T}, Τ::T, self::WaveMixing) where {T <: Real}

LEGACY Calculates the strong field Window from energies, oscillator strengths, and the probe laser amplitude as input. The inputs must be 1D Array, i.e. vectors. The signal is calculated merely from the dipole moments. Use with Excited State Dynamics.

Arguments

probe: probe pulse energy.

Τ: pulse length.

WaveMixing: composite containing the λ parameter. λ is the pump laser amplitude.

Returns

Strong field SE Window, matrix.

Description

This function estimates the strong field (window) parameter under the approximations of the Doorway-Window representation. The corresponding formula (see Ref. [1] formula XX) of the stimulated emission (SE) reads:

\[\]

Cite

[1] XX et.al. J. Chem. Phys.XXXX, X, XXXX

See also Δgsb_strong, gsb_strong, Δse_strong, and esa_strong.

Missing docstring.

Missing docstring for se_wind_2d. Check Documenter's build log for details.

Missing docstring.

Missing docstring for se_door_r2d. Check Documenter's build log for details.

Missing docstring.

Missing docstring for se_door_nr2d. Check Documenter's build log for details.

WaveMixings.se_fluorescence! — Function

se_fluorescence!(self::WaveMixing, probe::AbstractMatrix{T}, Τ::T) where {T <: Real}

Calculates the fluorescence signal from energies and oscillator strengths as input. The inputs must be 2D Array, i.e. Matrix. The signal is calculated from the spectrum energy in the frequency domain. For the available methods check the function spectrum_energy_fourier. Use with Excited State Dynamics.

Arguments

WaveMixing: composite containing oscillators, method and energies parameters.

probe: probe pulse energies.

Τ: pulse length.

Returns

Calculates fluorescence window and updates the WaveMixing composite (SE) with it. (Matrix with n-probe columns, one column has all sums for one probe pulse energy).

Description

This function estimates the Window (window) parameter under the following approximations:

Evolutions of electronic coherences in the DW operators are evaluated within the short-time classical Condon approximation.

T-evolutions of electronic populations, are replaced by evolutions along trajectories.

The quantum thermal distribution in the electronic ground state is substituted by the corresponding phase-space Wigner distribution.

The trace over vibrational degrees of freedom is replaced by stochastic sampling of the initial conditions.

The corresponding fluorescence formula (see Ref. [1] formula 22) of the stimulated emission (SE) reads:

\[S_{\alpha}^{TRF}(T,\omega_{t}) \sim \omega_{pu}\omega_{t}^{3} \left\langle D^{TFR}(\boldsymbol{R}_{g},\boldsymbol{P}_{g})W_{\mathrm{I}}^{TFR}(\omega_{t},\boldsymbol{R}(T),\boldsymbol{P}(T))\right\rangle\]

Cite

[1] Xu et.al. J. Chem. Phys.2024, 160, 104109

See also se, and se_strong_fluorescence!.

WaveMixings.Δse — Function

Δse(self::WaveMixing, frame::Integer=1, )

Calculates the Δ SE from energies and oscillator strengths as input. The inputs must be 1D Array, i.e. vectors. The signal is calculated merely from the dipole moments. Use with Excited State Dynamics.

Arguments

WaveMixing: composite containing oscillators, method and energies parameters.

frame: frame number in the non-adiabatic trajectory. The default gives 1.

Returns

SE delta_door, single value.

Description

This function estimates the Doorway (delta_door) parameter under the assumsion that the pump and probe pulses are well temporally separated i.e. the time delay (T) between the pulses is longer than the pulse durations, $\tau > T_{a}$ (see Ref. [1] formula 52)

\[D_{0}(\omega_{pu},\boldsymbol{R}_{g},\boldsymbol{P}_{g}) = \left| (\boldsymbol{\mu}_{ge}(\boldsymbol{R}_{g}) \right|^{2} \rho_{g}^{\textrm {Wig}}(\boldsymbol{R}_{g},\boldsymbol{P}_{g})\]

Cite

[1] Gelin et.al. J.Chem.TheoryComput.2021, 17, 2394−2408

See also Δse, and Δse_strong.

WaveMixings.Δse_lgc — Function

Δse_lgc(self::WaveMixing, frame::Integer=1, )

LEGACY Calculates the Δ SE from energies and oscillator strengths as input. The inputs must be 1D Array, i.e. vectors. The signal is calculated merely from the dipole moments. Use with Excited State Dynamics.

Arguments

WaveMixing: composite containing oscillators, method and energies parameters.

frame: frame number in the non-adiabatic trajectory. The default gives 1.

Returns

SE delta_door, single value.

Description

This function estimates the Doorway (delta_door) parameter under the assumsion that the pump and probe pulses are well temporally separated i.e. the time delay (T) between the pulses is longer than the pulse durations, $\tau > T_{a}$ (see Ref. [1] formula 52)

\[D_{0}(\omega_{pu},\boldsymbol{R}_{g},\boldsymbol{P}_{g}) = \left| (\boldsymbol{\mu}_{ge}(\boldsymbol{R}_{g}) \right|^{2} \rho_{g}^{\textrm {Wig}}(\boldsymbol{R}_{g},\boldsymbol{P}_{g})\]

Cite

[1] Gelin et.al. J.Chem.TheoryComput.2021, 17, 2394−2408

See also Δse, and Δse_strong.

WaveMixings.Δse_strong_lgc — Function

Δse_strong_lgc(self::WaveMixing, frame::Integer=1)

LEGACY Calculates the Δ strong field signal from energies, oscillator strengths, and the pump laser amplitude as input. The inputs must be 1D Array, i.e. vectors. The signal is calculated merely from the dipole moments. Use with Excited State Dynamics.

Arguments

WaveMixing: composite containing the λ parameter. λ is the pump laser amplitude.

frame: frame number in the non-adiabatic trajectory. The default gives 1.

Returns

Strong field SE delta_door, single value.

ESA Signal

WaveMixings.esa — Function

esa(probe::AbstractMatrix{T}, Τ::T, self::WaveMixing) where {T <: Real}

Calculates the signal from energies and oscillator strengths as input. The inputs must be 2D Array, i.e. Matrix. The signal is calculated from the spectrum energy in the frequency domain. For the available methods check the function spectrum_energy_fourier. Use with Excited State Absorption Dynamics.

Arguments

probe: probe pulse energies.

WaveMixing: composite containing oscillators, method and energies parameters.

Τ: pulse length.

Returns

Calculates window and updates the WaveMixing composite (ESA) with it. (Matrix with n-probe columns, one column has all sums for one probe pulse energy).

Description

This function estimates the Window (window) parameter under the following approximations:

Evolutions of electronic coherences in the DW operators are evaluated within the short-time classical Condon approximation.

T-evolutions of electronic populations, are replaced by evolutions along trajectories.

The quantum thermal distribution in the electronic ground state is substituted by the corresponding phase-space Wigner distribution.

The trace over vibrational degrees of freedom is replaced by stochastic sampling of the initial conditions.

The corresponding Window formula (see Ref. [1] formula 60) of the excited-state absorption (ESA) reads:

\[W_{II}^{int}(\omega_{pr},\boldsymbol{R}_{e},\boldsymbol{P}_{e}) = \sum_{f}^{ } E_{pr}^{2}(\omega_{pr}-U_{fe(T)}(\boldsymbol{R}_{e})) \left| (\boldsymbol{\mu}_{e(T)t}(\boldsymbol{R}_{e}) \right|^{2} \]

Cite

[1] Gelin et.al. J.Chem.TheoryComput.2021, 17, 2394−2408

See also se, and se_strong.

WaveMixings.esa_dispersion — Function

esa_dispersion(probe::T, dispersion::AbstractMatrix{T}, Τ::T,
               self::WaveMixing, dephase::T=0.01) where {T <: Real}

Calculates the dispersed signal (eq. 64 in Gelin JCTC, 2394. 2021) from energies and oscillator strengths as input. The inputs must be 2D Array, i.e. Matrix. The signal is calculated from the spectrum energy in the frequency domain. For the available methods check the function spectrum_energy_fourier. Use with Excited State Absorption Dynamics.

Arguments

probe: a probe pulse energy.

dispersion: array of dispersion energies.

Τ: pulse length.

WaveMixing: composite containing oscillators, method and energies parameters.

dephase: dephasing parameter. the default gives 66 fs.

Returns

Calculates dispersed window and updates the WaveMixing composite (ESA) with it. (Matrix with n-probe columns, one column has all sums for one probe pulse energy).

Description

This function estimates the dispersed Window (window) transient-absorption pump-probe signal, which derives from performing the Fourier transform to the integral pump-probe signal (see esa) with respect to $\tau_{t}$. The dispersed pump-probe signal is sensitive to electronic dephasing ($\lambda$), a parameter of this function.

The corresponding dispersed Window formula (see Ref. [1] formula 64) of the excited-state absorption (ESA) reads:

\[W_{\mathrm{II}}^{dis}(\omega_{pr},\omega,\boldsymbol{R}_{e}(T),\boldsymbol{P}_{e}(T)) = E_{pr}^{2}(\omega-\omega_{pr}) \sum_{f}\frac{\nu}{\nu^{2}+(\omega-\omega_{fe(T)}(\boldsymbol{R}_{e}(T)))^{2}} |\mu_{e(T)f}(\boldsymbol{R}_{e}(T))|^{2}\]

Cite

[1] Gelin et.al. J.Chem.TheoryComput.2021, 17, 2394−2408

See also gsb_dispersion and se_dispersion.

WaveMixings.esa_lgc — Function

esa_lgc(probe:AbstractMatrix{T}, Τ::T, self::WaveMixing) where {T <: Real}

LEGACY Calculates the signal from energies and oscillator strengths as input. The inputs must be 2D Array, i.e. Matrix. The signal is calculated from the spectrum energy in the frequency domain. For the available methods check the function spectrum_energy_fourier. Use with Excited State Absorption Dynamics.

Arguments

probe: probe pulse energies.

WaveMixing: composite containing oscillators, method and energies parameters.

Τ: pulse length.

Returns

Calculates window and updates the WaveMixing composite (ESA) with it. (Matrix with n-probe columns, one column has all sums for one probe pulse energy).

Description

This function estimates the Window (window) parameter under the following approximations:

Evolutions of electronic coherences in the DW operators are evaluated within the short-time classical Condon approximation.

T-evolutions of electronic populations, are replaced by evolutions along trajectories.

The quantum thermal distribution in the electronic ground state is substituted by the corresponding phase-space Wigner distribution.

The trace over vibrational degrees of freedom is replaced by stochastic sampling of the initial conditions.

The corresponding Window formula (see Ref. [1] formula 60) of the excited-state absorption (ESA) reads:

\[W_{II}^{int}(\omega_{pr},\boldsymbol{R}_{e},\boldsymbol{P}_{e}) = \sum_{f}^{ } E_{pr}^{2}(\omega_{pr}-U_{fe(T)}(\boldsymbol{R}_{e})) \left| (\boldsymbol{\mu}_{e(T)t}(\boldsymbol{R}_{e}) \right|^{2} \]

Cite

[1] Gelin et.al. J.Chem.TheoryComput.2021, 17, 2394−2408

See also se, and se_strong.

WaveMixings.esa_strong — Function

esa_strong(probe::AbstractMatrix{T}, Τ::T, self::WaveMixing) where {T <: Real}

Calculates the strong field Window from energies, oscillator strengths, and the probe laser amplitude as input. The inputs must be 1D Array, i.e. vectors. The signal is calculated merely from the dipole moments. Use with Excited State Absorption Dynamics.

Arguments

probe: probe pulse energy.

Τ: pulse length.

WaveMixing: composite containing the λ parameter. λ is the pump laser amplitude.

Returns

Strong field ESA Window, matrix.

Description

This function estimates the strong field (window) parameter under the approximations of the Doorway-Window representation. The corresponding formula (see Ref. [1] formula XX) of the excited state absorption (ESA) reads:

\[\]

Cite

[1] XX et.al. J. Chem. Phys.XXXX, X, XXXX

See also Δgsb_strong, gsb_strong, Δse_strong, and se_strong.

Constructors

WaveMixings.fs2toau2 — Function

fs2toau2(femto::AbstractFloat)

Converts a time value given in femtoseconds squared, to atomic units squared.

Arguments

-femto: a single numerical value.

Returns

Time in squared atomic units.

WaveMixings.fs3toau3 — Function

fs3toau3(femto::AbstractFloat)

Converts a time value given in femtoseconds cubic, to atomic units cubic.

Arguments

-femto: a single numerical value.

Returns

Time in cubic atomic units.

WaveMixings.make_nth_order_dispersion — Function

make_nth_order_dispersion(time::AbstractFloat, order::Integer)

Converts the pulse duration given in fs^n to au^n.

Arguments

time: pulse duration in fs^n.

n: dispersion order.

Returns

Pulse duration in au^n.

WaveMixings.make_pulse_duration — Function

make_pulse_duration(time::Real)

Converts the pulse duration given in fs to au.

Arguments

time: pulse duration in fs.

Returns

Pulse duration in au.

WaveMixings.make_pump_amplitude — Function

make_pump_amplitude(amplitude::Real)

Converts the pump amplitude given in eV to Hartree.

Arguments

amplitude: pump amplitude.

Returns

Pump amplitude in Hartree unit.

WaveMixings.make_pump_freq — Function

make_pump_freq(freq::Real)

Converts the pump frequency given in eV to Hartree.

Arguments

freq: pump frequency.

Returns

Pump frequency in Hartree unit.

WaveMixings.make_probe_freq — Function

make_probe_freq(start, stop, steps::Integer)

Creates a vector from start to stop at an interval step. This function ensures that the vector is of type Float64, thus speeding calculations.

Arguments

start: starting value of the frequency.

stop: ending value of the frequency.

steps: frequency intervals to compute.

Returns

Vector (1D array) with all the probe frequency values.

WaveMixings.make_door_window! — Function

make_door_window!(self::WaveMixing)

Arguments

WaveMixing: composite containing window, door and spectral_signal parameters.

Returns

Doorway-Window signal (spectra). In general this is the pump-probe spectra for given trajectory.

make_door_window!(esa::WaveMixing, se::WaveMixing)

Arguments

esa: composite containing window, and spectral_signal parameters.

se: composite containing door and spectral_signal parameters.

Returns

Doorway-Window signal (spectra) for the WaveMixing composite (ESA). In general this is the pump-probe spectra for given trajectory.

Misc.

WaveMixings.pkginfo — Method

pkginfo(io::IO=stdout)

Returns

Information on WaveMixings, dependencies, and system informations.

WaveMixings.cite — Method

WaveMixings.cite(io::IO = stdout)

Command line output of citation information and bibtex generator for WaveMixings.jl.

The outputs will be something like the following:

using WaveMixings
WaveMixings.pkginfo()

                              ___           
__          __ __  __        / _ \_/\       
\ \        / /|  \/  |  ___  \/ \___/___    
 \ \  /\  / / | \  / | / _ \_/\     / _ \_/\
  \ \/  \/ /  | |\/| | \/ \___/     \/ \___/
   \  /\  /   | |  | |
    \/  \/    |_|  |_|

Julia framework for computing time-resolved
nonlinear electronic spectra from on-the-fly
quasi-classical trajectories.
Keywords: transient absorption, pump-probe, TA PP.
≡≡≡≡≡≡≡≡≡≡≡≡≡≡≡≡≡≡≡≡≡≡≡≡≡≡≡≡≡≡≡≡≡≡≡≡≡≡≡≡≡≡≡≡≡≡≡≡≡≡≡≡
Repository      : codeberg.org/apolionl/WaveMixings.jl
Lead  developer : Dr. Luis Vasquez
Other developers: (Alphabetically)
                  Prof. Dr. Maxim Gelin
                  Prof. Dr. Zhenggang Lan
                  Dr. Chen Lipeng
                  Dr. Sebastian Pios
                  Prof. Dr. Ke-Wei Sun
Copyright © Gelin's group 2023 and later.

Package information:
====================================
Julia              Ver. 1.11.2
WaveMixings        Ver. 0.9.0
DelimitedFiles     Ver. 1.9.1
LinearAlgebra      Ver. 1.11.0

System information:
====================================
OS       : Linux (x86_64-pc-linux-gnu)
CPU      : 64 × Intel(R) Xeon(R) Gold 6426Y
Memory   : 1007.556 GB
WORD_SIZE: 64
LIBM     : libopenlibm
LLVM     : libLLVM-16.0.6 (ORCJIT, sapphirerapids)
BLAS     : libopenblas64_.so (ilp64)
BLAS CPU : 1 cores
Threads  : 1 (on 64 virtual cores)

+--------------------------------------------------------+
| Please cite WaveMixings.jl in your publication if used |
+--------------------------------------------------------+
CITATION.bib file contains the nesessary references of Wavemixings.jl.
Alternatively, one can use `WaveMixings.cite()`.

WaveMixings.print_logo — Method

print_logo(io::IO=stdout)

Print the Logo of WaveMixings package

The output will be something like the following:

using WaveMixings
WaveMixings.print_logo()

                              ___           
__          __ __  __        / _ \_/\       
\ \        / /|  \/  |  ___  \/ \___/___    
 \ \  /\  / / | \  / | / _ \_/\     / _ \_/\
  \ \/  \/ /  | |\/| | \/ \___/     \/ \___/
   \  /\  /   | |  | |
    \/  \/    |_|  |_|

Deprecated